First commit
Counting time from trigger to echo fall instead of active echo interval because it's simpler. But it's not following the datasheet, should be fixed in future versions.
Currently pin numbers are hard-coded and only two sonars are supported. They should be configurable through module parameters.
A bit buggy... For some reason weird huge numbers are returned when no object is in front of the sonar. In the example program (example.c), we handle it limiting the distance (if (packet0.distance > 600 * 1000 * 58) packet0.distance = 600 * 1000 * 58;).